backend/app/requirements.txt
aiobotocore==2.3.2
aiohttp==3.8.1
aioitertools==0.10.0
aiosignal==1.2.0
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
botocore==1.24.21
charset-normalizer==2.0.12
click==8.1.3
fastapi==0.78.0
frozenlist==1.3.0
h11==0.13.0
idna==3.3
jmespath==1.0.0
multidict==6.0.2
pydantic==1.9.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
wrapt==1.14.1
yarl==1.7.2
